+++ src/identity_mapper_service/__init__.py
@@ -1,6 +1,10 @@
 """IdentityMapper Host Service."""
 
-from identity_mapper_service.registry import ProviderRegistry, UnknownProviderError
+from identity_mapper_service.registry import (
+    ProviderRegistry,
+    UnknownProviderError,
+    UnknownTargetMapperError,
+)
 from identity_mapper_service.responses import (
     AuditResponse,
     HealthResponse,
@@ -15,4 +19,5 @@ __all__ = [
     "ProviderRegistry",
     "ProvidersResponse",
     "UnknownProviderError",
+    "UnknownTargetMapperError",
 ]